Extract SDK transport and storage internals
